Rename rev-parse --abbrev to --short.
The usage of rev-parse to serve as a flag/option parser
for git-whatchanged and other commands have serious limitation
that the flags cannot be something that is supported by
rev-parse itself, and it cannot worked around easily.  Since
this is rarely used "poor-man's describe", rename the option for
now as an easier workaround.

Signed-off-by: Junio C Hamano <junkio@cox.net>