rev-list --boundary: fix re-injecting boundary commits.
Marco reported that

	$ git rev-list --boundary --topo-order --parents 0eb4847..d54bbed

misses these two boundary commits.

        a550d33647e0be0b0a5ec42e0d3c09d18668c70d
        d6dec8099ee1d047d776ff814af5c119b6e38e6e

Indeed, we can see that gitk shows these two commits at the
bottom, because the --boundary code failed to output them.

The code did not check to avoid pushing the same uninteresting
commit twice to the result list.  I am not sure why this fixes
the reported problem, but this seems to fix it.

Signed-off-by: Junio C Hamano <junkio@cox.net>