gitk: Make File->Update work properly again
If a view is selected it will now just update that view.

Also fixed a few other things - if you switch away from a view while
gitk is still reading it in, then switch back, gitk will re-read it
from scratch.  We now re-read the references when switching views.
If something was selected before a view change, and we need to read
in the new view, we now select the previously-selected commit when
we come across it.

Fixed a bug in setting of rowrangelist plus a couple of other minor
things.

Signed-off-by: Paul Mackerras <paulus@samba.org>